experiments/rag-dynamic-db-v3/pyproject.toml
[tool.poetry]
name = "gedig-rag-v3"
version = "1.0.0"
description = "geDIG-based Self-Growing RAG System for Academic Publication"
authors = ["InsightSpike Research Team"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = ">=2.0.0"
transformers = ">=4.30.0"
sentence-transformers = ">=2.2.0"
networkx = ">=3.0"
numpy = ">=1.24.0"
pandas = ">=2.0.0"
scikit-learn = ">=1.3.0"
matplotlib = ">=3.7.0"
seaborn = ">=0.12.0"
plotly = ">=5.15.0"
pydantic = ">=2.0.0"
pyyaml = ">=6.0"
datasets = ">=2.14.0"
evaluate = ">=0.4.0"
rouge-score = ">=0.1.2"
nltk = ">=3.8"
spacy = ">=3.6.0"
tqdm = ">=4.65.0"
wandb = ">=0.15.0"
hydra-core = ">=1.3.0"
omegaconf = ">=2.3.0"
jupyterlab = ">=4.0.0"
ipywidgets = ">=8.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-cov = ">=4.1.0"
black = ">=23.7.0"
isort = ">=5.12.0"
flake8 = ">=6.0.0"
mypy = ">=1.5.0"
pre-commit = ">=3.3.0"
sphinx = ">=7.1.0"
sphinx-rtd-theme = ">=1.3.0"

[tool.poetry.group.paper.dependencies]
arxiv = ">=1.4.0"
bibtexparser = ">=1.4.0"
texsoup = ">=0.3.1"
pylatex = ">=1.4.1"

[tool.poetry.group.visualization.dependencies]
dash = ">=2.14.0"
dash-bootstrap-components = ">=1.4.0"
kaleido = ">=0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "longterm: Long-term experiment tests",
    "ablation: Ablation study tests",
    "slow: Slow tests (> 30s)",
    "gpu: Tests requiring GPU"
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "networkx.*",
    "sentence_transformers.*",
    "transformers.*",
    "datasets.*",
    "evaluate.*",
    "rouge_score.*",
    "wandb.*"
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
    "*/conftest.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod"
]
show_missing = true
precision = 2

[tool.poetry.scripts]
gedig-rag = "src.cli.main:main"
run-longterm = "src.experiments.longterm_runner:main"
run-ablation = "src.experiments.ablation_runner:main"
generate-figures = "src.visualization.paper_figure_generator:main"
generate-paper = "src.paper.paper_draft_generator:main"
quality-check = "src.quality.experiment_quality_checker:main"